requirements.txt
python-telegram-bot[job-queue]==21.7
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0